% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-calibration-engine.R
\name{apply_gtfs_speeds_to_ssfs}
\alias{apply_gtfs_speeds_to_ssfs}
\title{Speeds calibration engine}
\usage{
apply_gtfs_speeds_to_ssfs(
  gtfs,
  ssfs,
  max_date = NULL,
  buffer_dist = 10,
  dist_factor = 0.5,
  stop_time = 10,
  osrm_speed_adj_factor = 0.72,
  accepted_route_types = c(0, 3)
)
}
\arguments{
\item{gtfs}{An object of class 'gtfs'. Must include a shapes table.}

\item{ssfs}{A ssfs list}

\item{max_date}{A date within the range of gtfs$calendar$end_date representing the maximum of a 7 day range used to build the reference speed matrix from the reference GTFS. Leave as NULL to use the last 7 days of the reference GTFS}

\item{buffer_dist}{A distance in meters used to define the radius of interstop speed matrix points. Defaults to 10.}

\item{dist_factor}{A value between 0.1 and 0.9 used to ensure that interstops are only applied speeds from reference interstops of a similar length. Defaults to 0.5, which means that for a given interstop being calibrated, reference interstops with a distance of 50\% to 150\% can be used to calibrate.}

\item{stop_time}{An integer in seconds, representing the amount of time added per stop made for runtimes calculated using OSRM}

\item{osrm_speed_adj_factor}{A coefficient used to adjust output OSRM runtimes to make them more representative of bus runtimes.}

\item{accepted_route_types}{Route types that can be used to build the reference speed matrix. By default, 0 (tramways) and 3 (buses).}
}
\value{
A ssfs list
}
\description{
Reads a reference GTFS and a target SSFS and returns that SSFS with speeds adjusted
to those documented in the GTFS based on shapes and stop_times data.
}
\examples{
# Calibrate Railway City Transit redesign SSFS speeds using reference GTFS
ssfs_calibrated <- apply_gtfs_speeds_to_ssfs(gtfs = gtfs_rct, ssfs = ssfs_rct2)
}
