% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-stops-builders.R
\name{build_stop_row}
\alias{build_stop_row}
\title{Build a single stop row in the list}
\usage{
build_stop_row(stop)
}
\arguments{
\item{stop}{A one-row data.frame with stop_id and stop_name}
}
\description{
Build a single stop row in the list
}
