% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croquis-package.R
\docType{package}
\name{croquis-package}
\alias{croquis-package}
\title{croquis: Transit Sketch Planning with GTFS}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for rapidly creating detailed and realistic transit networks and schedules in GTFS format, including a front-end Shiny app called Croquis. Introduces the Simplified Speed and Frequency Specification (ssfs), a data structure that simplifies GTFS creation and editing. Includes functions for conversion between SSFS and GTFS formats. Includes a calibration engine which adjusts SSFS speeds based on a reference GTFS. Output GTFS are compatible with travel demand and accessibility analysis tools such as r5r.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/julian-city/croquis}
  \item Report bugs at \url{https://github.com/julian-city/croquis/issues}
}

}
\author{
\strong{Maintainer}: Julian Villafuerte Diaz \email{contact@julian.city} (\href{https://orcid.org/0000-0001-6543-7105}{ORCID})

}
\keyword{internal}
