% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companions.R
\name{generate_tdrh}
\alias{generate_tdrh}
\title{Generate trips, distance and runtime by hour (TDRH)}
\usage{
generate_tdrh(ssfs, id_type = c("route_id", "itin_id"), id, service)
}
\arguments{
\item{ssfs}{A list of class SSFS}

\item{id_type}{Either "route_id" or "itin_id"}

\item{id}{A character vector of one or more route_ids or itin_ids}

\item{service}{An individual string or vector representing one or several service_ids}
}
\value{
A tibble
}
\description{
Internal function that generates a data frame that can be used to calculate service cost
for specified routes or individual route itineraries in terms of service hours and
service kilometers.
}
\keyword{internal}
