% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtfs_rct}
\alias{gtfs_rct}
\title{Railway City Transit (St. Thomas, Ontario) GTFS}
\format{
A GTFS
}
\source{
Created by the package author based on public GTFS data.
}
\usage{
gtfs_rct
}
\description{
A sample GTFS representing the Railway City Transit network and schedule
in St. Thomas, Ontario.
}
\keyword{datasets}
