% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companions.R
\name{gtfs_remove_routes}
\alias{gtfs_remove_routes}
\title{Remove routes from GTFS}
\usage{
gtfs_remove_routes(gtfs, remove_routes)
}
\arguments{
\item{gtfs}{An object of class 'gtfs'}

\item{remove_routes}{A character vector of the route_ids that you wish to remove from the GTFS}
}
\value{
An object of class 'gtfs'
}
\description{
Remove routes from GTFS
}
\examples{
\dontrun{
# Import GTFS
gtfs <- gtfstools::read_gtfs("path/to/gtfs.zip")

# filter to remove specified routes
gtfs_filtered <- gtfs_remove_routes(gtfs,remove_routes = c("route_3","route_4"))
}
}
