% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companions.R
\name{gtfs_retain_routes}
\alias{gtfs_retain_routes}
\title{Retain routes from GTFS}
\usage{
gtfs_retain_routes(gtfs, retain_routes)
}
\arguments{
\item{gtfs}{An object of class 'gtfs'}

\item{retain_routes}{A character vector indicating the route_ids you wish to retain in the GTFS}
}
\value{
An object of class 'gtfs'
}
\description{
Retain routes from GTFS
}
\examples{
\dontrun{
# Import GTFS
gtfs <- gtfstools::read_gtfs("path/to/gtfs.zip")

# filter to retain specified routes
gtfs_filtered <- gtfs_retain_routes(gtfs,retain_routes = c("route_1","route_2"))
}
}
