% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ssfs_rct2}
\alias{ssfs_rct2}
\title{Railway City Transit (St. Thomas, Ontario) redesign scenario SSFS}
\format{
An ssfs object (a named list of 8 tables).
}
\source{
Created by the package author using Croquis
}
\usage{
ssfs_rct2
}
\description{
A sample SSFS representing a redesign of the Railway City Transit (St. Thomas, Ontario) network
}
\keyword{datasets}
