% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companions.R
\name{ssfs_subset}
\alias{ssfs_subset}
\title{Subset SSFS}
\usage{
ssfs_subset(
  ssfs,
  subset_id,
  operation = c("retain", "remove"),
  id_type = c("itin_id", "route_id")
)
}
\arguments{
\item{ssfs}{An input SSFS}

\item{subset_id}{The set of itin_ids or route_ids that you would like to remove or retain}

\item{operation}{Specify whether you would like the output ssfs to retain (default) or remove the data associated with subset_id}

\item{id_type}{Specify whether subset_id is itin_id (default) or route_id}
}
\value{
A SSFS
}
\description{
A handy function that outputs a subset of an input ssfs.
You can specify whether you want to remove or retain a based on a set of itin_ids or route_ids.
}
\examples{
#Create a subset SSFS that only includes data pertaining to route 160 of the mileend network
ssfs_160 <- ssfs_subset(mileend,subset_id="160",id_type="route_id")

#create a subset SSFS that excludes secondary itin_ids for various routes in of the mileend network
ssfs_mileend_clean <- ssfs_subset(mileend,subset_id=c("160_0_2","161_1_2"),operation="remove")
}
