% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{translink}
\alias{translink}
\title{Translink Vancouver network}
\format{
An ssfs object (a named list of 8 tables).
}
\source{
Created by the package author based on public Translink GTFS data.
}
\usage{
translink
}
\description{
A sample SSFS object representing the Translink Vancouver
public transit network.
}
\keyword{datasets}
