% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ttc_subway}
\alias{ttc_subway}
\title{TTC Subway sample network}
\format{
An ssfs object (a named list of 8 tables).
}
\source{
Created by the package author based on public GTFS data.
}
\usage{
ttc_subway
}
\description{
A sample SSFS object representing the Toronto Transit Commission
subway network.
}
\keyword{datasets}
