% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-calibration-engine.R
\name{apply_gtfs_speeds_to_ssfs}
\alias{apply_gtfs_speeds_to_ssfs}
\title{Speeds calibration engine}
\usage{
apply_gtfs_speeds_to_ssfs(
  gtfs,
  ssfs,
  max_date,
  buffer_dist = 10,
  dist_factor = 0.5,
  stop_time = 10,
  osrm_speed_adj_factor = 0.72
)
}
\arguments{
\item{gtfs}{An object of class 'gtfs'}

\item{ssfs}{A ssfs list}

\item{max_date}{A date representing the maximum date range of services retained from the reference GTFS}

\item{buffer_dist}{A distance in meters used to define the radius of interstop speed matrix points. Defaults to 10.}

\item{dist_factor}{A value between 0.1 and 0.9 used to ensure that interstops are only applied speeds from reference interstops of a similar length. Defaults to 0.5, which means that for a given interstop being calibrated, reference interstops with a distance of 50\% to 150\% can be used to calibrate.}

\item{stop_time}{An integer in seconds, representing the amount of time added per stop made for runtimes calculated using OSRM}

\item{osrm_speed_adj_factor}{A coefficient used to adjust output OSRM runtimes to make them more representative of bus runtimes.}
}
\value{
A ssfs list
}
\description{
Speeds calibration engine
}
\examples{
\dontrun{
# Create reference GTFS
gtfs <- gtfstools::read_gtfs("path/to/gtfs.zip")

# Import reference ssfs
ssfs <- readRDS("path/to/ssfs.rds")

# Apply the reference GTFS speeds to ssfs
ssfs_calibrated <- apply_gtfs_speeds_to_ssfs(gtfs, ssfs)

# With custom parameters
ssfs_calibrated <- apply_gtfs_speeds_to_ssfs(
  gtfs,
  ssfs,
  buffer_dist = 15,
  dist_factor = 0.4,
  stop_time = 12,
  osrm_speed_adj_factor = 0.70
)
}
}
