% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration-intermediaries.R
\name{apply_interstop_matrix_to_ssfs}
\alias{apply_interstop_matrix_to_ssfs}
\title{Apply interstop matrixx to ssfs}
\usage{
apply_interstop_matrix_to_ssfs(
  ssfs,
  interstop_matrices_by_service,
  buffer_dist = 10,
  dist_factor = 0.5,
  stop_time = 10,
  osrm_speed_adj_factor = 0.72
)
}
\arguments{
\item{ssfs}{A ssfs list}

\item{interstop_matrices_by_service}{An interstop speed matrix table}

\item{buffer_dist}{A distance in meters used to define the radius of interstop speed matrix points. Defaults to 10.}

\item{dist_factor}{A value between 0.1 and 0.9 used to ensure that interstops are only applied speeds from reference interstops of a similar length. Defaults to 0.5, which means that for a given interstop being calibrated, reference interstops with a distance of 50\% to 150\% can be used to calibrate.}

\item{stop_time}{An integer in seconds, representing the amount of time added per stop made for runtimes calculated using OSRM}

\item{osrm_speed_adj_factor}{A coefficient used to adjust output OSRM runtimes to make them more representative of bus runtimes.}
}
\value{
A ssfs list
}
\description{
Apply interstop matrixx to ssfs
}
