% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companions.R
\name{generate_service_cost}
\alias{generate_service_cost}
\title{Generate total daily costs in service hours and service kilometers}
\usage{
generate_service_cost(ssfs, id_type = c("route_id", "itin_id"), id, service)
}
\arguments{
\item{ssfs}{A list of class SSFS}

\item{id_type}{Either "route_id" or "itin_id"}

\item{id}{A character vector of one or more route_ids or itin_ids}

\item{service}{An individual string or vector representing one or several service_ids}
}
\value{
A tibble
}
\description{
Outputs a small tibble of costs for the specified service(s) and itin_id(s) or route_id(s)
}
\examples{
# Calculate weekday daily service kilometers and service hours of the 99 B-Line in Vancouver
b_line_route_id <- translink$routes |> dplyr::filter(route_short_name=="099") |> dplyr::pull(route_id)
generate_service_cost(ssfs=translink,id_type="route_id",id=b_line_route_id,service="mon-fri")
}
