% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companions.R
\name{generate_tdrh}
\alias{generate_tdrh}
\title{Generate trips, distance and runtime by hour (TDRH)}
\usage{
generate_tdrh(ssfs, id_type = c("route_id", "itin_id"), id, service)
}
\arguments{
\item{ssfs}{A list of class SSFS}

\item{id_type}{Either "route_id" or "itin_id"}

\item{id}{A character vector of one or more route_ids or itin_ids}

\item{service}{An individual string or vector representing one or several service_ids}
}
\description{
Function that generates a tibble that can be used to calculate service cost
for specified routes or individual route itineraries in terms of service hours and
service kilometers.
}
\examples{
# Generate table for the 99 B line to view runtimes by hour for all itin ids
b_line_route_id <- translink$routes |> dplyr::filter(route_short_name=="099") |> dplyr::pull(route_id)
generate_tdrh(ssfs=translink, id_type="route_id", id = b_line_route_id, service="mon-fri")
}
