% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration-intermediaries.R
\name{gtfs_to_interstop_matrix}
\alias{gtfs_to_interstop_matrix}
\title{Turn GTFS into interstop speed matrix}
\usage{
gtfs_to_interstop_matrix(gtfs, ssfs, max_date)
}
\arguments{
\item{gtfs}{An object with class 'gtfs'}

\item{ssfs}{A ssfs list}

\item{max_date}{A date representing the maximum date range of services retained from the reference GTFS}
}
\value{
An interstop speed matrix table
}
\description{
Turn GTFS into interstop speed matrix
}
\examples{
\dontrun{
# Create interstop speed matrix
# For now, specifying a max date within the gtfs is mandatory
# SSFS must be supplied to evaluate service structure
# First, create an interstop speed matrix from a reference GTFS
gtfs <- gtfstools::read_gtfs("path/to/gtfs.zip")
interstop_matrix <- gtfs_to_interstop_matrix(
  gtfs,
  ssfs,
  max_date = as.Date("2026-03-31")
)
}
}
