% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_to_ssfs.R
\name{gtfs_to_ssfs}
\alias{gtfs_to_ssfs}
\title{Convert a GTFS to a SSFS}
\usage{
gtfs_to_ssfs(
  gtfs,
  routes = NULL,
  max_date = NULL,
  routing_server = c("Valhalla", "OSRM")
)
}
\arguments{
\item{gtfs}{An object of class 'gtfs'. Must contain required tables and calendar table.}

\item{routes}{A character vector of the route id(s) you wish to convert to ssfs. Leave as NULL to convert all routes to SSFS}

\item{max_date}{A date within the range of gtfs$calendar$end_date representing the maximum of a 7 day range used to build the SSFS. Leave as NULL to use the last 7 days specified in gtfs$calendar to build the SSFS}

\item{routing_server}{Routing server used to draw shapes in the case where none are provided in the input GTFS}
}
\value{
A SSFS list
}
\description{
Reads a GTFS object and transforms it into a Simplified Schedule
and Frequency Structure (SSFS), extracting route geometries,
stop sequences, and service patterns into a more editable format
for sketch planning workflows.
}
\examples{
\donttest{
# First, create a GTFS from the sample Ligne Jaune network
gtfs <- ssfs_to_gtfs(stm_metro)

# Convert the GTFS back to SSFS
ssfs <- gtfs_to_ssfs(gtfs)

# Convert specific routes only
ssfs <- gtfs_to_ssfs(stm_metro, routes = c("1","2"))
}
}
