% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_to_ssfs.R
\name{gtfs_to_ssfs}
\alias{gtfs_to_ssfs}
\title{Convert a GTFS to a SSFS}
\usage{
gtfs_to_ssfs(gtfs, routes = NULL, max_date = NULL)
}
\arguments{
\item{gtfs}{An object of class 'gtfs'. Must contain required tables and calendar table.}

\item{routes}{A character vector of the route id(s) you wish to convert to ssfs. Leave as NULL to convert all routes to SSFS}

\item{max_date}{A date within the range of gtfs$calendar$end_date representing the maximum of a 7 day range used to build the SSFS. Leave as NULL to use the last 7 days specified in gtfs$calendar to build the SSFS}
}
\value{
A SSFS list
}
\description{
Convert a GTFS to a SSFS
}
\examples{
\dontrun{
# Load a GTFS file using gtfstools
gtfs <- gtfstools::read_gtfs("path/to/gtfs.zip")

# Convert all routes to SSFS.
ssfs <- gtfs_to_ssfs(gtfs)

# Convert specific routes only
ssfs <- gtfs_to_ssfs(gtfs, routes = c("route_1", "route_2"))

# Specify maximum date
# By default, max_date is NULL and the gtfs_to_ssfs looks at the last 7 days of the GTFS service period
# max_date indicates the last day of the 7 day period desired for use to build the SSFS
ssfs <- gtfs_to_ssfs(gtfs, max_date = as.Date("2026-03-31"))
}
}
