% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_to_ssfs.R
\name{gtfs_to_ssfs}
\alias{gtfs_to_ssfs}
\title{Convert a GTFS to a SSFS}
\usage{
gtfs_to_ssfs(gtfs, routes = NULL, max_date = NULL)
}
\arguments{
\item{gtfs}{An object of class 'gtfs'. Must contain required tables and calendar table.}

\item{routes}{A character vector of the route id(s) you wish to convert to ssfs. Leave as NULL to convert all routes to SSFS}

\item{max_date}{A date within the range of gtfs$calendar$end_date representing the maximum of a 7 day range used to build the SSFS. Leave as NULL to use the last 7 days specified in gtfs$calendar to build the SSFS}
}
\value{
A SSFS list
}
\description{
Convert a GTFS to a SSFS
}
