% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ligne_jaune}
\alias{ligne_jaune}
\title{STM Ligne Jaune sample network}
\format{
An ssfs object (a named list of 8 tables).
}
\source{
Created by the package author based on public STM GTFS data.
}
\usage{
ligne_jaune
}
\description{
A sample SSFS (Simplified Frequencies and Speeds Structure) object representing
the STM Ligne Jaune (Yellow Line) metro service in Montreal.
This is the simplest sample network and is
useful for getting started with Croquis.
}
\keyword{datasets}
