% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mileend}
\alias{mileend}
\title{STM Mile-End bus sample network}
\format{
An ssfs object (a named list of 8 tables).
}
\source{
Created by the package author based on public STM GTFS data.
}
\usage{
mileend
}
\description{
A sample SSFS object representing a subset of the STM bus network
in the Mile-End neighbourhood of Montreal.
}
\keyword{datasets}
