% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssfs_to_gtfs.R
\name{ssfs_to_gtfs}
\alias{ssfs_to_gtfs}
\title{Convert a SSFS to a GTFS}
\usage{
ssfs_to_gtfs(ssfs, dist_traveled = FALSE)
}
\arguments{
\item{ssfs}{An ssfs list}

\item{dist_traveled}{When TRUE, adds shape_dist_traveled field to shapes and stop_times tables in output GTFS}
}
\value{
an object of class 'gtfs'
}
\description{
Convert a SSFS to a GTFS
}
\examples{
# Convert the sample Ligne Jaune SSFS to GTFS
gtfs <- ssfs_to_gtfs(ligne_jaune)

\donttest{
# Include shape_dist_traveled (increases processing time)
gtfs_with_dist <- ssfs_to_gtfs(ligne_jaune, dist_traveled = TRUE)
}
}
