% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssfs_to_gtfs.R
\name{ssfs_to_gtfs}
\alias{ssfs_to_gtfs}
\title{Convert a SSFS to a GTFS}
\usage{
ssfs_to_gtfs(ssfs, dist_traveled = FALSE)
}
\arguments{
\item{ssfs}{An ssfs list}

\item{dist_traveled}{When TRUE, adds shape_dist_traveled field to shapes and stop_times tables in output GTFS}
}
\value{
an object of class 'gtfs'
}
\description{
Convert a SSFS to a GTFS
}
\examples{
\dontrun{
# Create a GTFS based on a SSFS
gtfs <- ssfs_to_gtfs(ssfs)
# Create shape_dist_traveled fields in GTFS stop_times and shapes tables
# This increases processing time
gtfs <- ssfs_to_gtfs(ssfs, dist_traveled = TRUE)
}
}
