% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stm_metro}
\alias{stm_metro}
\title{STM Metro sample network}
\format{
An ssfs object (a named list of 8 tables).
}
\source{
Created by the package author based on public STM GTFS data.
}
\usage{
stm_metro
}
\description{
A sample SSFS object representing the full STM metro network in Montreal.
}
\keyword{datasets}
