% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{trip_dep_generator}
\alias{trip_dep_generator}
\title{Generate trip departure times for a time range}
\usage{
trip_dep_generator(ssfs, first_dep, last_dep, itin_id_i, service_id_i)
}
\arguments{
\item{ssfs}{A list of class SSFS}

\item{first_dep}{A string indicating first departure time in HH:MM:SS format}

\item{last_dep}{A string indicating last departure time in HH:MM:SS format}

\item{itin_id_i}{A string indicating a specific itin_id}

\item{service_id_i}{A string indicating a specific service_id}
}
\value{
A vector of strings of trip departure times in HH:MM:SS format
}
\description{
Internal function that generates trip departure times for a specified range,
based on a first departure and a last departure (for example the bounds of a service window)
and based on the hsh values of the ssfs, for a specific itin_id and service_id.
Used within ssfs_to_gtfs() as well as within the cost calculator function
}
\keyword{internal}
