% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssfs-class.R
\name{validate_ssfs}
\alias{validate_ssfs}
\title{Validate an ssfs object}
\usage{
validate_ssfs(x)
}
\arguments{
\item{x}{An object to validate, typically created by \code{\link[=new_ssfs]{new_ssfs()}} or
\code{\link[=gtfs_to_ssfs]{gtfs_to_ssfs()}}.}
}
\value{
\code{x}, invisibly, if validation passes.
}
\description{
Checks that an ssfs object contains the 8 required tables and that each
table contains its required columns. Returns the object invisibly if
valid; throws an informative error otherwise.
}
\examples{
\dontrun{
# Run validation. Console messages will indicate missing tables, missing vectors (columns), or incorrect class / data type
validate_ssfs(ssfs)
}
}
